Sorry i had to make a name change duo to the "#" symbol in name.
I apologize for the inconvenience, but I had to change the name due to the "#" symbol causing issues with the new auto-updater.

To remove the "#" symbol for the Lua settings, you can easily open your profile through a text editor.

Unfortunately, the previous version of "1# PIP Girl" will no longer receive updates due to the name change. Despite my best efforts to find a solution, I was unable to do so without causing further issues. However, the new and improved "1 PIP Girl" will now receive updates as usual. I sincerely apologize for any inconvenience this may have caused, and I am committed to ensuring that similar issues do not occur in the future.

To get the latest version of "1 PIP Girl", please delete the old Lua file and download the new version using the ScriptManager with my GitHub link or you can do this directly from Github or through Discord.