improvements on kick blacklist on join
In one of the recent updates, a new feature was introduced. Now, if a person on the blacklist is identified as a stand user, they will not be immediately kicked. Instead, they will be put in "kick-on-attack" mode. This means that if they initiate an attack, whether it's a genuine threat or a false positive, they will be promptly kicked from the session.