Done some stuff
the new import has some issues
so i fixed some on the old one .

the new import:
has issues with checking if the requested RID is in the history and couses false non exsistence in history sometimes, couse of issue is know and i try to fix it.

old importer:
abuses more the api , so some player wont be added to the history ,and bcs of that cant be blocked.