:hypers:
More Auto Accept Warnings.

Added Session Claimer Size in > Session:
Select the Size of a Session u want to claim.
This Value can be saved in a Profile!^^
(!) Size 31-32 is very rare to reach, so its only use would be filling the Player History.

Added Session Claimer in > Session:
Finds a Session with Selcted Size.
Checks if the host is not a Modder or Friend.
Claims the Host if all clear and next as host.
Else looks for a better place to stay.
Admin Bailing & Auto Accept Warning included.