Update 1 PIP Girl.lua
Updated Auto Accept Warning and Auto Skip Cutscene.

if u find any location that "Auto Skip Cutscene" or "Auto Accept Warning" is breaking stuff or should not skip , pls DM me the location of the issue or go in the settings of my lua and copy the location and send me it in DM's with info about the issue.